style(test): replace int with explicit-width integer types
Replace ambiguous int/unsigned int with stdint types throughout
test shims and spec files:
- uint16_t for test counters (testCount, testPasses)
- uint32_t for line numbers in bddtest_test
- bool for connect() return and bddtest_test result parameter
- int16_t for available()/read()/peek() (sentinel -1 still fits)
- uint8_t for array indices, loop counters, and small length values
- int8_t for MQTT state (range -4..5)
- Add <cstdint> include to BDDTest.h and BDDTest.cpp

+++ test/_shims/src/IPAddress.h
@@ -52,8 +52,8 @@ public:
     bool operator==(const uint8_t* addr);
 
     // Overloaded index operator to allow getting and setting individual octets of the address
-    uint8_t operator[](int index) const { return _address[index]; };
-    uint8_t& operator[](int index) { return _address[index]; };
+    uint8_t operator[](uint8_t index) const { return _address[index]; };
+    uint8_t& operator[](uint8_t index) { return _address[index]; };
 
     // Overloaded copy operators to allow initialisation of IPAddress objects from other types
     IPAddress& operator=(const uint8_t *address);
